app/requirements.txt
celery==5.2.6
Flask==2.1.2
#Flask-PyMongo==2.3.0
gunicorn==20.1.0
opencv-contrib-python==4.6.0.66
#opencv-python-headless==4.6.0.66
#opencv-python==4.6.0.66
pytest==7.2.1
python-dotenv==1.0.0
requests==2.20.1
redis
pymongo==3.11
pillow
numpy